replace image loading with manifest
instead of loading HTMLImageElements to get image dimensions, use the dims already in the manifest.json file. fixes a bug where browser isn't recognizing the image in the cache and is loading them twice (once for dims, and again in the SVGImageElement href). doesn't seem to be a way to convert from HTMLImageElement -> SVGImageElement without using the expensive data URL.